functions/requirements.txt
requests==2.28.1
google-cloud-bigquery==3.3.2
